Fix edge case when the image block's caption field is empty on page load
