+++ packages/block-library/src/image/test/use-open-image-media-editor-modal.js
@@ -15,7 +15,11 @@ import { useRegistry, useSelect } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import { useOpenImageMediaEditorModal } from '../use-open-image-media-editor-modal';
+import {
+	getImageBlockMetadataFromAttachment,
+	getSyncedImageBlockAttributes,
+	useOpenImageMediaEditorModal,
+} from '../use-open-image-media-editor-modal';
 
 const mockOpenMediaEditorModalKey = 'openMediaEditorModal';
 
@@ -96,7 +100,10 @@ describe( 'useOpenImageMediaEditorModal', () => {
 		};
 		const registry = createRegistry( {
 			getEntityRecord: () => originalAttachment,
-			resolveGetEntityRecord: () => updatedAttachment,
+			resolveGetEntityRecord: ( kind, name, attachmentId, query ) =>
+				query?.context === 'edit'
+					? updatedAttachment
+					: originalAttachment,
 		} );
 		useRegistry.mockReturnValue( registry );
 		const setAttributes = jest.fn();
@@ -131,6 +138,10 @@ describe( 'useOpenImageMediaEditorModal', () => {
 		expect( registry.actions.invalidateResolution ).toHaveBeenCalledWith(
 			'getEntityRecord',
 			[ 'postType', 'attachment', 1 ]
+		);
+		expect( registry.actions.invalidateResolution ).toHaveBeenCalledWith(
+			'getEntityRecord',
+			[ 'postType', 'attachment', 1, { context: 'edit' } ]
 		);
 		expect( registry.actions.invalidateResolution ).toHaveBeenCalledWith(
 			'getEntityRecord',
@@ -186,7 +197,8 @@ describe( 'useOpenImageMediaEditorModal', () => {
 			1,
 			'postType',
 			'attachment',
-			1
+			1,
+			{ context: 'edit' }
 		);
 		expect( openMediaEditorModal ).toHaveBeenCalledWith( {
 			id: 1,
@@ -195,6 +207,128 @@ describe( 'useOpenImageMediaEditorModal', () => {
 		expect( setAttributes ).toHaveBeenCalledWith( {
 			alt: 'Updated alt',
 			caption: 'Updated caption',
+		} );
+	} );
+
+	it( 'resolves original raw attachment metadata before opening the modal when the block has no caption', async () => {
+		const originalAttachment = {
+			id: 1,
+			alt_text: '',
+			caption: { raw: 'Existing attachment caption' },
+		};
+		const updatedAttachment = {
+			id: 1,
+			alt_text: '',
+			caption: { raw: 'Updated attachment caption' },
+		};
+		const resolveGetEntityRecord = jest
+			.fn()
+			.mockResolvedValueOnce( originalAttachment )
+			.mockResolvedValueOnce( updatedAttachment );
+		const registry = createRegistry( {
+			resolveGetEntityRecord,
+		} );
+		useRegistry.mockReturnValue( registry );
+		const setAttributes = jest.fn();
+		const openMediaEditorModal = jest.fn();
+		useMediaEditorModalSetting( openMediaEditorModal );
+		const { result } = renderHook( () =>
+			useOpenImageMediaEditorModal( {
+				attributes: {
+					id: 1,
+					url: 'original.jpg',
+					alt: '',
+					caption: undefined,
+				},
+				setAttributes,
+			} )
+		);
+
+		await act( async () => {
+			await result.current();
+		} );
+		await act( async () => {
+			await openMediaEditorModal.mock.calls[ 0 ][ 0 ].onUpdate( {
+				id: 1,
+				url: 'updated.jpg',
+			} );
+		} );
+
+		expect( resolveGetEntityRecord ).toHaveBeenNthCalledWith(
+			1,
+			'postType',
+			'attachment',
+			1,
+			{ context: 'edit' }
+		);
+		expect( openMediaEditorModal ).toHaveBeenCalledWith( {
+			id: 1,
+			onUpdate: expect.any( Function ),
+		} );
+		expect( setAttributes ).toHaveBeenCalledWith( {
+			caption: 'Updated attachment caption',
+		} );
+	} );
+
+	it( 'resolves original raw attachment metadata before opening the modal when the cached record has only a rendered caption', async () => {
+		const cachedAttachment = {
+			id: 1,
+			alt_text: '',
+			caption: { rendered: '<p>Existing attachment caption</p>\n' },
+		};
+		const originalAttachment = {
+			id: 1,
+			alt_text: '',
+			caption: { raw: 'Existing attachment caption' },
+		};
+		const updatedAttachment = {
+			id: 1,
+			alt_text: '',
+			caption: { raw: 'Updated attachment caption' },
+		};
+		const resolveGetEntityRecord = jest
+			.fn()
+			.mockResolvedValueOnce( originalAttachment )
+			.mockResolvedValueOnce( updatedAttachment );
+		const registry = createRegistry( {
+			getEntityRecord: () => cachedAttachment,
+			resolveGetEntityRecord,
+		} );
+		useRegistry.mockReturnValue( registry );
+		const setAttributes = jest.fn();
+		const openMediaEditorModal = jest.fn();
+		useMediaEditorModalSetting( openMediaEditorModal );
+		const { result } = renderHook( () =>
+			useOpenImageMediaEditorModal( {
+				attributes: {
+					id: 1,
+					url: 'original.jpg',
+					alt: '',
+					caption: undefined,
+				},
+				setAttributes,
+			} )
+		);
+
+		await act( async () => {
+			await result.current();
+		} );
+		await act( async () => {
+			await openMediaEditorModal.mock.calls[ 0 ][ 0 ].onUpdate( {
+				id: 1,
+				url: 'updated.jpg',
+			} );
+		} );
+
+		expect( resolveGetEntityRecord ).toHaveBeenNthCalledWith(
+			1,
+			'postType',
+			'attachment',
+			1,
+			{ context: 'edit' }
+		);
+		expect( setAttributes ).toHaveBeenCalledWith( {
+			caption: 'Updated attachment caption',
 		} );
 	} );
 
@@ -356,13 +490,22 @@ describe( 'useOpenImageMediaEditorModal', () => {
 	} );
 
 	it( 'syncs metadata from an empty block when the original attachment is not cached', async () => {
+		const originalAttachment = {
+			id: 1,
+			alt_text: '',
+			caption: { raw: '' },
+		};
 		const updatedAttachment = {
 			id: 1,
 			alt_text: 'Updated alt',
 			caption: { raw: 'Updated caption' },
 		};
+		const resolveGetEntityRecord = jest
+			.fn()
+			.mockResolvedValueOnce( originalAttachment )
+			.mockResolvedValueOnce( updatedAttachment );
 		const registry = createRegistry( {
-			resolveGetEntityRecord: () => updatedAttachment,
+			resolveGetEntityRecord,
 		} );
 		useRegistry.mockReturnValue( registry );
 		const setAttributes = jest.fn();
@@ -393,6 +536,52 @@ describe( 'useOpenImageMediaEditorModal', () => {
 		expect( setAttributes ).toHaveBeenCalledWith( {
 			alt: 'Updated alt',
 			caption: 'Updated caption',
+		} );
+	} );
+
+	it( 'does not sync a field that was not changed in the modal', async () => {
+		const originalAttachment = {
+			id: 1,
+			alt_text: 'Original alt',
+			caption: { raw: 'Existing caption' },
+		};
+		const updatedAttachment = {
+			id: 1,
+			alt_text: 'Updated alt',
+			caption: { raw: 'Existing caption' },
+		};
+		const registry = createRegistry( {
+			getEntityRecord: () => originalAttachment,
+			resolveGetEntityRecord: () => updatedAttachment,
+		} );
+		useRegistry.mockReturnValue( registry );
+		const setAttributes = jest.fn();
+		const openMediaEditorModal = jest.fn();
+		useMediaEditorModalSetting( openMediaEditorModal );
+		const { result } = renderHook( () =>
+			useOpenImageMediaEditorModal( {
+				attributes: {
+					id: 1,
+					url: 'original.jpg',
+					alt: 'Original alt',
+					caption: undefined,
+				},
+				setAttributes,
+			} )
+		);
+
+		await act( async () => {
+			await result.current();
+		} );
+		await act( async () => {
+			await openMediaEditorModal.mock.calls[ 0 ][ 0 ].onUpdate( {
+				id: 1,
+				url: 'original.jpg',
+			} );
+		} );
+
+		expect( setAttributes ).toHaveBeenCalledWith( {
+			alt: 'Updated alt',
 		} );
 	} );
 
@@ -502,5 +691,235 @@ describe( 'useOpenImageMediaEditorModal', () => {
 			alt: 'Attachment alt',
 			caption: 'Attachment caption',
 		} );
+	} );
+} );
+
+describe( 'getImageBlockMetadataFromAttachment', () => {
+	it( 'normalizes attachment metadata to image block attributes', () => {
+		expect(
+			getImageBlockMetadataFromAttachment( {
+				alt_text: 'Alt text',
+				caption: { raw: 'First line\nSecond line' },
+			} )
+		).toEqual( {
+			alt: 'Alt text',
+			caption: 'First line<br>Second line',
+		} );
+	} );
+
+	it( 'does not use rendered captions when raw captions are unavailable', () => {
+		expect(
+			getImageBlockMetadataFromAttachment( {
+				alt_text: 'Alt text',
+				caption: { rendered: '<p>Rendered caption</p>\n' },
+			} )
+		).toEqual( {
+			alt: 'Alt text',
+			caption: undefined,
+		} );
+	} );
+
+	it( 'preserves paragraph markup in raw captions', () => {
+		expect(
+			getImageBlockMetadataFromAttachment( {
+				caption: { raw: '<p>Raw caption</p>' },
+			} ).caption
+		).toBe( '<p>Raw caption</p>' );
+	} );
+
+	it( 'does not fall back to rendered captions when raw captions are empty', () => {
+		expect(
+			getImageBlockMetadataFromAttachment( {
+				caption: {
+					raw: '',
+					rendered: '<p>Rendered caption</p>\n',
+				},
+			} ).caption
+		).toBe( '' );
+	} );
+
+	it( 'returns an unknown caption when only rendered empty caption markup is available', () => {
+		expect(
+			getImageBlockMetadataFromAttachment( {
+				caption: {
+					rendered: '<p class="attachment"><br></p>\n',
+				},
+			} ).caption
+		).toBe( undefined );
+	} );
+} );
+
+describe( 'getSyncedImageBlockAttributes', () => {
+	it( 'syncs updated attachment metadata when block metadata was not customized', () => {
+		expect(
+			getSyncedImageBlockAttributes(
+				{
+					alt: 'Original alt',
+					caption: 'Original caption',
+				},
+				{
+					alt_text: 'Original alt',
+					caption: { raw: 'Original caption' },
+				},
+				{
+					alt_text: 'Updated alt',
+					caption: { raw: 'Updated caption' },
+				}
+			)
+		).toEqual( {
+			alt: 'Updated alt',
+			caption: 'Updated caption',
+		} );
+	} );
+
+	it( 'does not overwrite custom block alt text', () => {
+		expect(
+			getSyncedImageBlockAttributes(
+				{
+					alt: 'Custom alt',
+					caption: 'Original caption',
+				},
+				{
+					alt_text: 'Original alt',
+					caption: { raw: 'Original caption' },
+				},
+				{
+					alt_text: 'Updated alt',
+					caption: { raw: 'Updated caption' },
+				}
+			)
+		).toEqual( {
+			caption: 'Updated caption',
+		} );
+	} );
+
+	it( 'does not overwrite custom block captions', () => {
+		expect(
+			getSyncedImageBlockAttributes(
+				{
+					alt: 'Original alt',
+					caption: 'Custom caption',
+				},
+				{
+					alt_text: 'Original alt',
+					caption: { raw: 'Original caption' },
+				},
+				{
+					alt_text: 'Updated alt',
+					caption: { raw: 'Updated caption' },
+				}
+			)
+		).toEqual( {
+			alt: 'Updated alt',
+		} );
+	} );
+
+	it( 'syncs newly added attachment metadata when original metadata was empty', () => {
+		expect(
+			getSyncedImageBlockAttributes(
+				{},
+				{
+					alt_text: '',
+					caption: { raw: '' },
+				},
+				{
+					alt_text: 'Updated alt',
+					caption: { raw: 'Updated\ncaption' },
+				}
+			)
+		).toEqual( {
+			alt: 'Updated alt',
+			caption: 'Updated<br>caption',
+		} );
+	} );
+
+	it( 'does not sync captions when the original raw attachment caption is unavailable', () => {
+		expect(
+			getSyncedImageBlockAttributes(
+				{},
+				{
+					caption: {
+						rendered: '<p>Original caption</p>\n',
+					},
+				},
+				{
+					caption: { raw: 'Updated caption' },
+				}
+			)
+		).toEqual( {} );
+	} );
+
+	it( 'syncs caption to a block with no caption when the original attachment has one', () => {
+		expect(
+			getSyncedImageBlockAttributes(
+				{
+					alt: '',
+					caption: '',
+				},
+				{
+					alt_text: '',
+					caption: { raw: 'Existing caption' },
+				},
+				{
+					alt_text: '',
+					caption: { raw: 'Updated caption' },
+				}
+			)
+		).toEqual( {
+			caption: 'Updated caption',
+		} );
+	} );
+
+	it( 'does not sync caption when block has a custom value differing from the original', () => {
+		expect(
+			getSyncedImageBlockAttributes(
+				{
+					alt: '',
+					caption: 'Custom caption',
+				},
+				{
+					alt_text: '',
+					caption: { raw: 'Original caption' },
+				},
+				{
+					alt_text: '',
+					caption: { raw: 'Updated caption' },
+				}
+			)
+		).toEqual( {} );
+	} );
+
+	it( 'clears captions when the updated attachment caption is empty', () => {
+		expect(
+			getSyncedImageBlockAttributes(
+				{
+					caption: 'Original caption',
+				},
+				{
+					caption: { raw: 'Original caption' },
+				},
+				{
+					caption: { raw: '' },
+				}
+			)
+		).toEqual( {
+			caption: undefined,
+		} );
+	} );
+
+	it( 'does not sync when the original attachment metadata is unknown', () => {
+		expect(
+			getSyncedImageBlockAttributes(
+				{
+					alt: '',
+					caption: '',
+				},
+				undefined,
+				{
+					alt_text: 'Updated alt',
+					caption: { raw: 'Updated caption' },
+				}
+			)
+		).toEqual( {} );
 	} );
 } );
