+++ packages/block-library/src/image/use-open-image-media-editor-modal.js
@@ -6,16 +6,107 @@ import {
 	privateApis as blockEditorPrivateApis,
 	store as blockEditorStore,
 } from '@wordpress/block-editor';
+import { __unstableStripHTML as stripHTML } from '@wordpress/dom';
 import { useRegistry, useSelect } from '@wordpress/data';
 import { useCallback, useEffect, useRef } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
-import { getSyncedImageBlockAttributes } from './utils';
 import { unlock } from '../lock-unlock';
 
+function normalizeImageBlockCaption( caption ) {
+	if ( typeof caption !== 'string' ) {
+		return '';
+	}
+
+	const textContent = stripHTML( caption ).trim();
+
+	if ( ! textContent ) {
+		return '';
+	}
+
+	return caption.replace( /\n/g, '<br>' );
+}
+
+function getAttachmentCaption( attachment ) {
+	const caption = attachment?.caption;
+
+	if ( typeof caption === 'string' ) {
+		return normalizeImageBlockCaption( caption );
+	}
+
+	if (
+		caption &&
+		typeof caption === 'object' &&
+		Object.hasOwn( caption, 'raw' )
+	) {
+		return normalizeImageBlockCaption( caption.raw );
+	}
+
+	return undefined;
+}
+
+export function getImageBlockMetadataFromAttachment( attachment ) {
+	return {
+		alt:
+			typeof attachment?.alt_text === 'string'
+				? attachment.alt_text
+				: attachment?.alt || '',
+		caption: getAttachmentCaption( attachment ),
+	};
+}
+
+function normalizeMetadataAttribute( value ) {
+	return value || '';
+}
+
+export function getSyncedImageBlockAttributes(
+	currentAttributes,
+	originalAttachment,
+	updatedAttachment
+) {
+	if ( ! originalAttachment || ! updatedAttachment ) {
+		return {};
+	}
+
+	const originalMetadata =
+		getImageBlockMetadataFromAttachment( originalAttachment );
+	const updatedMetadata =
+		getImageBlockMetadataFromAttachment( updatedAttachment );
+	const syncedAttributes = {};
+
+	const normalizedCurrentAlt = normalizeMetadataAttribute(
+		currentAttributes.alt
+	);
+	if (
+		originalMetadata.alt !== updatedMetadata.alt &&
+		( normalizedCurrentAlt === originalMetadata.alt ||
+			! normalizedCurrentAlt )
+	) {
+		syncedAttributes.alt = updatedMetadata.alt;
+	}
+
+	const normalizedCurrentCaption = normalizeMetadataAttribute(
+		currentAttributes.caption
+	);
+	if (
+		originalMetadata.caption !== undefined &&
+		updatedMetadata.caption !== undefined &&
+		originalMetadata.caption !== updatedMetadata.caption &&
+		( normalizedCurrentCaption === originalMetadata.caption ||
+			! normalizedCurrentCaption )
+	) {
+		syncedAttributes.caption = updatedMetadata.caption || undefined;
+	}
+
+	return syncedAttributes;
+}
+
 const { openMediaEditorModalKey } = unlock( blockEditorPrivateApis );
+// Caption sync needs `caption.raw`; view/default attachment records can contain
+// only rendered caption data or be tied to an in-flight stale resolution.
+const ATTACHMENT_EDIT_QUERY = { context: 'edit' };
 
 function getAttachmentFallbackForEmptyBlockMetadata( { alt, caption } ) {
 	const attachment = {};
@@ -24,11 +115,25 @@ function getAttachmentFallbackForEmptyBlockMetadata( { alt, caption } ) {
 		attachment.alt_text = '';
 	}
 
-	if ( ! caption ) {
+	if ( ! caption?.toString() ) {
 		attachment.caption = '';
 	}
 
 	return Object.keys( attachment ).length ? attachment : undefined;
+}
+
+function hasKnownAttachmentMetadata( attachment ) {
+	if ( ! attachment ) {
+		return false;
+	}
+
+	const hasKnownAlt =
+		typeof attachment.alt_text === 'string' ||
+		typeof attachment.alt === 'string';
+	const hasKnownCaption =
+		getImageBlockMetadataFromAttachment( attachment ).caption !== undefined;
+
+	return hasKnownAlt && hasKnownCaption;
 }
 
 export function useOpenImageMediaEditorModal( { attributes, setAttributes } ) {
@@ -42,12 +147,12 @@ export function useOpenImageMediaEditorModal( { attributes, setAttributes } ) {
 			select( blockEditorStore ).getSettings()[ openMediaEditorModalKey ],
 		[]
 	);
-	const blockMetadataRef = useRef( { alt, caption } );
+	const blockMetadataRef = useRef( { alt, caption: caption?.toString() } );
 	const mediaEditorMetadataBaselineRef = useRef();
 	const mediaEditorMetadataSyncRequestRef = useRef( 0 );
 
 	useEffect( () => {
-		blockMetadataRef.current = { alt, caption };
+		blockMetadataRef.current = { alt, caption: caption?.toString() };
 	}, [ alt, caption ] );
 
 	const getCachedAttachmentRecord = useCallback(
@@ -59,7 +164,14 @@ export function useOpenImageMediaEditorModal( { attributes, setAttributes } ) {
 					'postType',
 					'attachment',
 					attachmentId
-				) || getEntityRecord( 'postType', 'attachment', attachmentId )
+				) ||
+				getEntityRecord(
+					'postType',
+					'attachment',
+					attachmentId,
+					ATTACHMENT_EDIT_QUERY
+				) ||
+				getEntityRecord( 'postType', 'attachment', attachmentId )
 			);
 		},
 		[ registry ]
@@ -67,10 +179,22 @@ export function useOpenImageMediaEditorModal( { attributes, setAttributes } ) {
 
 	const resolveAttachmentRecord = useCallback(
 		async ( attachmentId ) => {
+			const resolveSelect = registry.resolveSelect( coreStore );
+
 			try {
-				return await registry
-					.resolveSelect( coreStore )
-					.getEntityRecord( 'postType', 'attachment', attachmentId );
+				return (
+					( await resolveSelect.getEntityRecord(
+						'postType',
+						'attachment',
+						attachmentId,
+						ATTACHMENT_EDIT_QUERY
+					) ) ||
+					( await resolveSelect.getEntityRecord(
+						'postType',
+						'attachment',
+						attachmentId
+					) )
+				);
 			} catch {
 				return undefined;
 			}
@@ -91,18 +215,18 @@ export function useOpenImageMediaEditorModal( { attributes, setAttributes } ) {
 				'postType',
 				'attachment',
 				attachmentId,
+				ATTACHMENT_EDIT_QUERY,
+			] );
+			invalidateResolution( 'getEntityRecord', [
+				'postType',
+				'attachment',
+				attachmentId,
 				{ context: 'view' },
 			] );
 
-			try {
-				return await registry
-					.resolveSelect( coreStore )
-					.getEntityRecord( 'postType', 'attachment', attachmentId );
-			} catch {
-				return undefined;
-			}
+			return resolveAttachmentRecord( attachmentId );
 		},
-		[ registry ]
+		[ registry, resolveAttachmentRecord ]
 	);
 
 	const handleMediaUpdate = useCallback(
@@ -165,15 +289,19 @@ export function useOpenImageMediaEditorModal( { attributes, setAttributes } ) {
 			getAttachmentFallbackForEmptyBlockMetadata(
 				blockMetadataRef.current
 			);
-		const needsOriginalAttachmentRecord =
-			! cachedAttachmentRecord &&
-			( blockMetadataRef.current.alt ||
-				blockMetadataRef.current.caption );
+		const resolvedAttachmentRecord = hasKnownAttachmentMetadata(
+			cachedAttachmentRecord
+		)
+			? undefined
+			: await resolveAttachmentRecord( id );
 
-		mediaEditorMetadataBaselineRef.current = needsOriginalAttachmentRecord
-			? ( await resolveAttachmentRecord( id ) ) ||
-			  fallbackAttachmentRecord
-			: cachedAttachmentRecord || fallbackAttachmentRecord;
+		mediaEditorMetadataBaselineRef.current =
+			resolvedAttachmentRecord ||
+			( hasKnownAttachmentMetadata( cachedAttachmentRecord )
+				? cachedAttachmentRecord
+				: fallbackAttachmentRecord ) ||
+			cachedAttachmentRecord;
+
 		openMediaEditorModal( {
 			id,
 			onUpdate: handleMediaUpdate,
