+++ packages/block-library/src/image/utils.js
@@ -1,9 +1,4 @@
 /**
- * WordPress dependencies
- */
-import { __unstableStripHTML as stripHTML } from '@wordpress/dom';
-
-/**
  * Internal dependencies
  */
 import { NEW_TAB_REL, ALLOWED_MEDIA_TYPES } from './constants';
@@ -101,92 +96,4 @@ export function isValidFileType( file ) {
 
 export function mediaPosition( { x, y } = { x: 0.5, y: 0.5 } ) {
 	return `${ Math.round( x * 100 ) }% ${ Math.round( y * 100 ) }%`;
-}
-
-function normalizeImageBlockCaption( caption ) {
-	if ( typeof caption !== 'string' ) {
-		return '';
-	}
-
-	const textContent = stripHTML( caption ).trim();
-
-	if ( ! textContent ) {
-		return '';
-	}
-
-	return caption.replace( /\n/g, '<br>' );
-}
-
-function getAttachmentCaption( attachment ) {
-	const caption = attachment?.caption;
-
-	if ( typeof caption === 'string' ) {
-		return normalizeImageBlockCaption( caption );
-	}
-
-	if (
-		caption &&
-		typeof caption === 'object' &&
-		Object.hasOwn( caption, 'raw' )
-	) {
-		return normalizeImageBlockCaption( caption.raw );
-	}
-
-	return undefined;
-}
-
-export function getImageBlockMetadataFromAttachment( attachment ) {
-	return {
-		alt:
-			typeof attachment?.alt_text === 'string'
-				? attachment.alt_text
-				: attachment?.alt || '',
-		caption: getAttachmentCaption( attachment ),
-	};
-}
-
-function normalizeMetadataAttribute( value ) {
-	return value || '';
-}
-
-export function getSyncedImageBlockAttributes(
-	currentAttributes,
-	originalAttachment,
-	updatedAttachment
-) {
-	if ( ! originalAttachment || ! updatedAttachment ) {
-		return {};
-	}
-
-	const originalMetadata =
-		getImageBlockMetadataFromAttachment( originalAttachment );
-	const updatedMetadata =
-		getImageBlockMetadataFromAttachment( updatedAttachment );
-	const syncedAttributes = {};
-
-	const normalizedCurrentAlt = normalizeMetadataAttribute(
-		currentAttributes.alt
-	);
-	if (
-		( normalizedCurrentAlt === originalMetadata.alt ||
-			! normalizedCurrentAlt ) &&
-		normalizedCurrentAlt !== updatedMetadata.alt
-	) {
-		syncedAttributes.alt = updatedMetadata.alt;
-	}
-
-	const normalizedCurrentCaption = normalizeMetadataAttribute(
-		currentAttributes.caption
-	);
-	if (
-		originalMetadata.caption !== undefined &&
-		updatedMetadata.caption !== undefined &&
-		( normalizedCurrentCaption === originalMetadata.caption ||
-			! normalizedCurrentCaption ) &&
-		normalizedCurrentCaption !== updatedMetadata.caption
-	) {
-		syncedAttributes.caption = updatedMetadata.caption || undefined;
-	}
-
-	return syncedAttributes;
 }
